authication

+++ backend/app/Http/Controllers/ExpenseController.php
@@ -6,19 +6,23 @@ use App\Models\Expense;
 use App\Models\Group;
 use Illuminate\Http\Request;
 use Illuminate\View\View;
+use Illuminate\Support\Facades\Auth;
 
 class ExpenseController extends Controller
 {
     public function index()
     {
-        $expenses = Expense::with('group')->get();
+        $user = Auth::user();
+
+        $expenses = $user->expenses()->with('group')->get();
         return response()->json($expenses);
         // return view('expenses.index', compact('expenses'));
     }
 
     public function create(): View
     {
-        $groups = Group::all();
+        $user = Auth::user();
+        $groups = Group::where('user_id', $user->id)->get();
         return view('expenses.create', compact('groups'));
     }
     public function store(Request $request)
@@ -29,7 +33,8 @@ class ExpenseController extends Controller
             'group_id' => 'required|exists:groups,id',
             'date' => 'required|date'
         ]);
-
+        $user = Auth::user();
+        $validated['user_id'] = $user->id; // Add user_id to the validated data
         $data = Expense::create($validated);
         $expense = Expense::with('group')->find($data->id);
         return response()->json([
@@ -40,16 +45,26 @@ class ExpenseController extends Controller
 
     public function show(Expense $expense): View
     {
+
+        if ($expense->user_id !== auth()->id) {
+            abort(403, 'Unauthorized action.');
+        };
         return view('expenses.show', compact('expense'));
     }
     public function edit(Expense $expense): View
     {
-        $groups = Group::all();
+        if ($expense->user_id !== auth()->id) {
+            abort(403, 'Unauthorized action.');
+        };
+        $groups = Group::where('user_id', auth()->id)->get();
         return view('expenses.edit', compact('expense', 'groups'));
     }
     public function update(Request $request, Expense $expense)
     {
-
+        $user = Auth::user();
+        if ($expense->user_id !== $user->id) {
+            abort(403, 'Unauthorized action.');
+        };
         $validated = $request->validate([
             'name' => 'required|max:255',
             'amount' => 'required',
@@ -69,6 +84,10 @@ class ExpenseController extends Controller
 
     public function destroy(Expense $expense)
     {
+        $user = Auth::user();
+        if ($expense->user_id !==  $user->id) {
+            abort(403, 'Unauthorized action.');
+        };
         $expense->delete();
 
         return response()->json([
